CDash warning solved
git-svn-id: https://svn.opendtect.org/od/trunk@34269 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/General/property.cc
@@ -355,7 +355,7 @@ float MathProperty::gtVal( Property::EvalOpts eo ) const
 	const Property* prop = inps_[iinp];
 	float val;
 	if ( !prop )
-	    val = form_.getConstVal( iinp );
+	    val = (float)form_.getConstVal( iinp );
 	else
 	{
 	    if ( prop == &xposprop )
